+++ include/aikido/planner/dart/ConfigurationToTSRPlanner.hpp
@@ -1,30 +1,34 @@
-#ifndef AIKIDO_PLANNER_CONFIGURATIONTOCONFIGURATIONPLANNER_HPP_
-#define AIKIDO_PLANNER_CONFIGURATIONTOCONFIGURATIONPLANNER_HPP_
+#ifndef AIKIDO_PLANNER_DART_CONFIGURATIONTOTSRPLANNER_HPP_
+#define AIKIDO_PLANNER_DART_CONFIGURATIONTOTSRPLANNER_HPP_
 
-#include "aikido/planner/ConfigurationToConfiguration.hpp"
-#include "aikido/planner/SingleProblemPlanner.hpp"
+#include "aikido/planner/dart/ConfigurationToTSR.hpp"
+#include "aikido/planner/dart/SingleProblemPlanner.hpp"
+#include "aikido/statespace/dart/MetaSkeletonStateSpace.hpp"
 #include "aikido/trajectory/Trajectory.hpp"
 
 namespace aikido {
 namespace planner {
+namespace dart {
 
-/// Base planner class for ConfigurationToConfiguration planning problem.
-class ConfigurationToConfigurationPlanner
-    : public SingleProblemPlanner<ConfigurationToConfigurationPlanner,
-                                  ConfigurationToConfiguration>
+/// Base planner class for ConfigurationToTSR planning problem.
+class ConfigurationToTSRPlanner
+    : public dart::SingleProblemPlanner<ConfigurationToTSRPlanner,
+                                        ConfigurationToTSR>
 {
 public:
   // Expose the implementation of Planner::plan(const Problem&, Result*) in
   // SingleProblemPlanner. Note that plan() of the base class takes Problem
   // while the virtual function defined in this class takes SolvableProblem,
-  // which is simply ConfigurationToConfiguration.
+  // which is simply ConfigurationToTSR.
   using SingleProblemPlanner::plan;
 
   /// Constructor
   ///
   /// \param[in] stateSpace State space that this planner associated with.
-  explicit ConfigurationToConfigurationPlanner(
-      statespace::ConstStateSpacePtr stateSpace);
+  /// \param[in] metaSkeleton MetaSkeleton to use for planning.
+  ConfigurationToTSRPlanner(
+      statespace::dart::ConstMetaSkeletonStateSpacePtr stateSpace,
+      ::dart::dynamics::MetaSkeletonPtr metaSkeleton);
 
   /// Solves \c problem returning the result to \c result.
   ///
@@ -36,7 +40,8 @@ public:
   // Note: SolvableProblem is defined in SingleProblemPlanner.
 };
 
+} // namespace dart
 } // namespace planner
 } // namespace aikido
 
-#endif // AIKIDO_PLANNER_CONFIGURATIONTOCONFIGURATIONPLANNER_HPP_
+#endif // AIKIDO_PLANNER_DART_CONFIGURATIONTOTSRPLANNER_HPP_