+++ src/planner/dart/ConfigurationToEndEffectorOffset.cpp
@@ -1,14 +1,15 @@
-#include "aikido/planner/ConfigurationToEndEffectorOffset.hpp"
+#include "aikido/planner/dart/ConfigurationToEndEffectorOffset.hpp"
 
 #include "aikido/constraint/Testable.hpp"
 
 namespace aikido {
 namespace planner {
+namespace dart {
 
 //==============================================================================
 ConfigurationToEndEffectorOffset::ConfigurationToEndEffectorOffset(
     statespace::dart::ConstMetaSkeletonStateSpacePtr stateSpace,
-    dart::dynamics::ConstBodyNodePtr endEffectorBodyNode,
+    ::dart::dynamics::ConstBodyNodePtr endEffectorBodyNode,
     const statespace::dart::MetaSkeletonStateSpace::State* startState,
     const Eigen::Vector3d& direction,
     const double signedDistance,
@@ -37,7 +38,7 @@ const std::string& ConfigurationToEndEffectorOffset::getStaticType()
 }
 
 //==============================================================================
-dart::dynamics::ConstBodyNodePtr
+::dart::dynamics::ConstBodyNodePtr
 ConfigurationToEndEffectorOffset::getEndEffectorBodyNode() const
 {
   return mEndEffectorBodyNode;
@@ -62,5 +63,6 @@ double ConfigurationToEndEffectorOffset::getDistance() const
   return mDistance;
 }
 
+} // namespace dart
 } // namespace planner
 } // namespace aikido